requirements.txt
addict
future
lmdb
numpy>=1.17
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
torch>=2.0
torchvision
tqdm
yapf
mmcv-full
mmedit
